requirements.txt
python-dotenv
loguru
Faker
faker-datasets
sqlalchemy